Address review comments: move ThumbnailContentProvider, rename ImageParsingDebugModule, update authority
