+++ app/src/main/java/org/oppia/android/app/application/dev/BUILD.bazel
@@ -10,7 +10,6 @@ kt_android_library(
         "DeveloperApplicationComponent.kt",
         "DeveloperBuildFlavorModule.kt",
         "DeveloperOppiaApplication.kt",
-        "ThumbnailContentProvider.kt",
     ],
     visibility = [
         "//:oppia_binary_visibility",
@@ -26,6 +25,7 @@ kt_android_library(
         "//domain/src/main/java/org/oppia/android/domain/platformparameter:debug_module",
         "//utility/src/main/java/org/oppia/android/util/logging/firebase:debug_module",
         "//utility/src/main/java/org/oppia/android/util/networking:debug_module",
-        "//utility/src/main/java/org/oppia/android/util/parser/image:dev_image_parsing_module",
+        "//app/src/main/java/org/oppia/android/app/utility/imageloading:thumbnail_content_provider",
+        "//utility/src/main/java/org/oppia/android/util/parser/image:image_parsing_debug_module",
     ],
 )
