+++ app/src/main/java/org/oppia/android/app/application/dev/DeveloperApplicationComponent.kt
@@ -61,7 +61,7 @@ import org.oppia.android.util.networking.NetworkConnectionDebugUtilModule
 import org.oppia.android.util.networking.NetworkConnectionUtilDebugModule
 import org.oppia.android.util.parser.html.HtmlParserEntityTypeModule
 import org.oppia.android.util.parser.image.GlideImageLoaderModule
-import org.oppia.android.util.parser.image.DevImageParsingModule
+import org.oppia.android.util.parser.image.ImageParsingDebugModule
 import org.oppia.android.util.system.OppiaClockModule
 import org.oppia.android.util.threading.DispatcherModule
 import javax.inject.Singleton
@@ -79,7 +79,7 @@ import javax.inject.Singleton
     MultipleChoiceInputModule::class, NumberWithUnitsRuleModule::class,
     NumericInputRuleModule::class, TextInputRuleModule::class, DragDropSortInputModule::class,
     InteractionsModule::class, GcsResourceModule::class, GlideImageLoaderModule::class,
-    DevImageParsingModule::class, HtmlParserEntityTypeModule::class, CachingModule::class,
+    ImageParsingDebugModule::class, HtmlParserEntityTypeModule::class, CachingModule::class,
     QuestionModule::class, DebugLogReportingModule::class, AccessibilityProdModule::class,
     ImageClickInputModule::class, LogStorageModule::class, IntentFactoryShimModule::class,
     ViewBindingShimModule::class,
