+++ app/src/main/java/org/oppia/android/app/utility/imageloading/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+  package="org.oppia.android">
+  <application>
+    <provider
+      android:name=".app.utility.imageloading.ThumbnailContentProvider"
+      android:authorities="org.oppia.android.provider.gcs"
+      android:exported="false" />
+  </application>
+</manifest>
